`aws` directory now `ons_aws` and package renamed too
